# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Comment:
    properties:
      author:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.CommentForm:
    properties:
      author:
        maxLength: 100
        minLength: 1
        type: string
      content:
        maxLength: 500
        minLength: 1
        type: string
      photo_id:
        type: integer
    required:
    - author
    - content
    - photo_id
    type: object
  models.File:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      id:
        type: integer
      mime_type:
        type: string
      oss_url:
        description: OSS访问URL
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
...
  utils.BusinessError:
    properties:
      bizCode:
        type: integer
      details: {}
      err: {}
      httpCode:
        type: integer
    type: object
  utils.ListResponse:
    properties:
      list: {}
      total:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample go-learn server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: go-learn API
  version: "1.0"
paths:
...
  /comments:
    post:
      consumes:
      - application/json
      description: 创建一条新的评论
      parameters:
      - description: 评论信息
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.CommentForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 创建评论
      tags:
      - comments
  /comments/{id}:
    delete:
      description: 根据ID删除评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 删除评论
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: 根据ID更新评论内容
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论信息
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.CommentForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 更新评论
      tags:
      - comments
  /files:
    get:
      description: 获取所有文件，支持分页
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取所有文件
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件到OSS
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 上传文件
      tags:
      - files
  /files/{id}:
    delete:
      description: 根据ID删除文件
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 删除文件
      tags:
      - files
    get:
      description: 根据ID获取文件详情
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取单个文件信息
      tags:
      - files
...
  /photos/{id}/comments:
    get:
      description: 根据照片ID获取所有评论，支持分页
      parameters:
      - description: 照片ID
        in: path
        name: id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取照片的所有评论
      tags:
      - comments
...
swagger: "2.0"